[COLAB-2918]: Use lombok to generate the client-side POJOs from a minimal POJO which only contains member variables
